+++ blogos-api/src/main/java/com/duan/blogos/api/blogger/BloggerStatisticsController.java
@@ -25,7 +25,7 @@ public class BloggerStatisticsController extends BaseBloggerController {
 
     @RequestMapping(method = RequestMethod.GET)
     public ResultModel<BloggerStatisticsDTO> get(HttpServletRequest request,
-                                                 @PathVariable Integer bloggerId) {
+                                                 @PathVariable Long bloggerId) {
 
         handleAccountCheck(bloggerId);
