+++ blogos-service-impl/src/main/java/com/duan/blogos/service/dao/blog/BlogDao.java
@@ -25,7 +25,7 @@ public interface BlogDao extends BaseDao<Blog> {
      * @return 查询到的文章
      * @see BlogStatusEnum
      */
-    List<Blog> listBlog(@Param("bloggerId") int bloggerId,
+    List<Blog> listBlog(@Param("bloggerId") Long bloggerId,
                         @Param("status") int status);
 
     /**
@@ -38,7 +38,7 @@ public interface BlogDao extends BaseDao<Blog> {
      * @return
      * @see BlogStatusEnum
      */
-    List<Blog> listBlogWithLimit(@Param("bloggerId") int bloggerId,
+    List<Blog> listBlogWithLimit(@Param("bloggerId") Long bloggerId,
                                  @Param("status") int status,
                                  @Param("offset") int offset,
                                  @Param("rows") int rows);
@@ -57,7 +57,7 @@ public interface BlogDao extends BaseDao<Blog> {
      * @param status    博文状态
      * @return 查询结果
      */
-    List<Blog> listAllCategoryAndLabel(@Param("bloggerId") int bloggerId, @Param("status") int status);
+    List<Blog> listAllCategoryAndLabel(@Param("bloggerId") Long bloggerId, @Param("status") int status);
 
     /**
      * 根据博文id查询博文
@@ -66,7 +66,7 @@ public interface BlogDao extends BaseDao<Blog> {
      * @param status 博文状态
      * @return 查询结果
      */
-    List<Blog> listBlogByBlogIds(@Param("ids") List<Integer> ids,
+    List<Blog> listBlogByBlogIds(@Param("ids") List<Long> ids,
                                  @Param("status") int status,
                                  @Param("offset") int offset,
                                  @Param("rows") int rows);
@@ -78,8 +78,8 @@ public interface BlogDao extends BaseDao<Blog> {
      * @param title     博文标题（同一博主的标题不能重复）
      * @return 查询结果
      */
-    Integer getBlogIdByUniqueKey(@Param("bloggerId") int bloggerId,
-                                 @Param("title") String title);
+    Long getBlogIdByUniqueKey(@Param("bloggerId") Long bloggerId,
+                              @Param("title") String title);
 
     /**
      * 通过博文id查询博文
@@ -87,7 +87,7 @@ public interface BlogDao extends BaseDao<Blog> {
      * @param blogId 博文id
      * @return 查询结果
      */
-    Blog getBlogById(int blogId);
+    Blog getBlogById(Long blogId);
 
     /**
      * 通过博文id查询博文id，该方法只为检查博文是否存在
@@ -95,7 +95,7 @@ public interface BlogDao extends BaseDao<Blog> {
      * @param blogId 博文id
      * @return 查询结果
      */
-    Integer getBlogIdById(int blogId);
+    Integer getBlogIdById(Long blogId);
 
     /**
      * 查询出指定博主的所有博文包含着的类别
@@ -103,7 +103,7 @@ public interface BlogDao extends BaseDao<Blog> {
      * @param bloggerId 博主id
      * @return 只查询了博文类别和博文id的集合
      */
-    List<Blog> listAllCategoryByBloggerId(int bloggerId);
+    List<Blog> listAllCategoryByBloggerId(Long bloggerId);
 
     /**
      * 查询出指定博主的所有博文包含的标签
@@ -111,7 +111,7 @@ public interface BlogDao extends BaseDao<Blog> {
      * @param bloggerId 博主id
      * @return 只查询了博文标签和博文id的结果集
      */
-    List<Blog> listAllLabelByBloggerId(int bloggerId);
+    List<Blog> listAllLabelByBloggerId(Long bloggerId);
 
     /**
      * 根据博主获得其所有博文的字数统计
@@ -119,7 +119,7 @@ public interface BlogDao extends BaseDao<Blog> {
      * @param bloggerId 博主id
      * @return 只有 wordCount 属性有值的结果
      */
-    List<Blog> listAllWordCountByBloggerId(@Param("bloggerId") int bloggerId,
+    List<Blog> listAllWordCountByBloggerId(@Param("bloggerId") Long bloggerId,
                                            @Param("state") int state);
 
     /**
@@ -129,8 +129,8 @@ public interface BlogDao extends BaseDao<Blog> {
      * @param categoryId 类别id
      * @return 数量
      */
-    Integer countBlogByCategory(@Param("bloggerId") int bloggerId,
-                                @Param("categoryId") int categoryId,
+    Integer countBlogByCategory(@Param("bloggerId") Long bloggerId,
+                                @Param("categoryId") Long categoryId,
                                 @Param("state") int state);
 
     /**
@@ -139,7 +139,7 @@ public interface BlogDao extends BaseDao<Blog> {
      * @param bloggerId 博主id
      * @return 数量
      */
-    Integer countBlogByBloggerId(@Param("bloggerId") int bloggerId,
+    Integer countBlogByBloggerId(@Param("bloggerId") Long bloggerId,
                                  @Param("state") int state);
 
     /**
@@ -149,7 +149,7 @@ public interface BlogDao extends BaseDao<Blog> {
      * @param format    md 或 html
      * @return 查询结果，format 为 md 时只查询 content_md，为 html 时只查询 content
      */
-    List<Blog> listAllByFormat(@Param("bloggerId") int bloggerId,
+    List<Blog> listAllByFormat(@Param("bloggerId") Long bloggerId,
                                @Param("format") int format);
 
     /**
@@ -158,5 +158,5 @@ public interface BlogDao extends BaseDao<Blog> {
      * @param bloggerId 博主id
      * @return 查询结果，只查询 id
      */
-    List<Blog> listAllIdByBloggerId(int bloggerId);
+    List<Blog> listAllIdByBloggerId(Long bloggerId);
 }
